extracted appcoins-services.json fields from the build.gradle.kts
